distro: move `newImageTypeFrom()` close to `imageType`
This commit just moves the existing `newImageTypeFrom()`
close to the `imageType` definition. The old place was
a separate file called `imagedefs.go` that (kinda) made
sense at some point but now the new*() should really
be close to the type definition.
